refactor(network): dissolve indexing import cycle
The Indexing/ResponseParser/ResponseParserLogs files formed a 3-node
import cycle: Indexing re-exported parseResponse+handleResponse from
ResponseParser (the cluster facade), while ResponseParser and
ResponseParserLogs imported request-meta primitives back from
Indexing.

Extract the request-meta extraction + capture-gate predicates
(extractRequestMeta, parseTextOrNull, shouldRecordResponse,
isUnsupportedUrl, isJsonContentType + their constants/types) verbatim
into a new leaf ResponsePrimitives.ts that depends only on the WK
registry, Debug, and playwright types. ResponseParser and
ResponseParserLogs now import the primitives from that leaf, removing
both back-edges into Indexing. Indexing re-exports every moved symbol
so its public facade stays byte-identical for downstream consumers.

Cycle gate: 2 cycles -> 1 (the 31-node core remains); coupling mass
34 -> 31. Baseline re-frozen; 270 unit tests pass.
